vfs7xxx: refactored power management interface
The old, split up and untimed pm-interface to cause FP Sensor OOO-errors (FP Sensor is out of order)
by not being able to boot the fingerprint sensors before the FPBAuthService tried to lock it.

This unexepcted behaviour (which seems to be limited to newer bootloaders) caused the already mentioned

    E bauth_FPBAuthService: FP Sensor is out of order 1, 0, 0, 0

error. This was solvable by either rebooting or killing the fingerprint-service. If this behaviour continues
even after this patch, a unconventional solution, crashing the fp-service when receiving the indicating error.

Change-Id: Idac4274f20590551585580d43f28be25fceb3972